+++ src/collective/multilingual/setuphandlers.py
@@ -11,11 +11,10 @@ except ImportError:
 
 
 @implementer(INonInstallable)
-class HiddenProfiles(object):
+class HiddenProfiles:
     def getNonInstallableProfiles(self):
         """Hide all profiles from site-creation and quickinstaller."""
         return [
-            "collective.multilingual:default",
             "collective.multilingual:uninstall",
         ]
 
@@ -23,9 +22,8 @@ class HiddenProfiles(object):
         """Hide the upgrades package from site-creation and quickinstaller.
 
         Our upgrades profiles are defined in the directory 'upgrades'.
-        Plone sees this is a separate product.
-        So instead of adding each new upgrade profile to the list of
-        non installable profiles above, we can mark the upgrades product
-        as non installable.
+        Plone sees this is a separate product. So instead of adding each
+        new upgrade profile to the list of non installable profiles
+        above, we can mark the upgrades product as non installable.
         """
         return ["collective.multilingual.upgrades"]